(refactor) rename interface jsx file to be capitalized
